+++ src/egon/data/airflow/entrypoints/create_postgis_extension.sql
@@ -1,3 +1,0 @@
-CREATE EXTENSION hstore;
-CREATE EXTENSION postgis;
-CREATE EXTENSION postgis_raster;