+++ src/egon/data/processing/substation/__init__.py
@@ -13,8 +13,11 @@ Base = declarative_base()
 class EgonEhvSubstation(Base):
     __tablename__ = 'egon_ehv_substation'
     __table_args__ = {'schema': 'grid'}
-    subst_id = Column(Integer, Sequence('ehv_id_seq'),
-        server_default=Sequence('ehv_id_seq').next_value(), primary_key=True)
+    subst_id = Column(Integer,
+        Sequence('egon_ehv_substation_subst_id_seq', schema='grid'),
+        server_default=
+            Sequence('egon_ehv_substation_subst_id_seq', schema='grid').next_value(),
+        primary_key=True)
     lon = Column(Float(53))
     lat = Column(Float(53))
     point = Column(Geometry('POINT', 4326), index=True)
@@ -35,8 +38,11 @@ class EgonEhvSubstation(Base):
 class EgonHvmvSubstation(Base):
     __tablename__ = 'egon_hvmv_substation'
     __table_args__ = {'schema': 'grid'}
-    subst_id = Column(Integer, Sequence('hvmv_id_seq'),
-        server_default=Sequence('hvmv_id_seq').next_value(), primary_key=True)
+    subst_id = Column(Integer,
+        Sequence('egon_hvmv_substation_subst_id_seq', schema='grid'),
+        server_default=
+            Sequence('egon_hvmv_substation_subst_id_seq', schema='grid').next_value(),
+        primary_key=True)
     lon = Column(Float(53))
     lat = Column(Float(53))
     point = Column(Geometry('POINT', 4326), index=True)
@@ -53,6 +59,29 @@ class EgonHvmvSubstation(Base):
     dbahn = Column(Text)
     status = Column(Integer)
 
+class EgonHvmvSubstationVoronoi(Base):
+    __tablename__ = 'egon_hvmv_substation_voronoi'
+    __table_args__ = {'schema': 'grid'}
+    id = Column(Integer,
+        Sequence('egon_hvmv_substation_voronoi_id_seq', schema='grid'),
+        server_default=
+            Sequence('egon_hvmv_substation_voronoi_id_seq', schema='grid').next_value(),
+        primary_key=True)
+    subst_id = Column(Integer)
+    geom = Column(Geometry('Multipolygon', 4326))
+
+
+class EgonEhvSubstationVoronoi(Base):
+    __tablename__ = 'egon_ehv_substation_voronoi'
+    __table_args__ = {'schema': 'grid'}
+    id = Column(Integer,
+        Sequence('egon_ehv_substation_voronoi_id_seq', schema='grid'),
+        server_default=
+            Sequence('egon_ehv_substation_voronoi_id_seq', schema='grid').next_value(),
+        primary_key=True)
+    subst_id = Column(Integer)
+    geom = Column(Geometry('Multipolygon', 4326))
+
 def create_tables():
     """Create tables for substation data
     Returns
@@ -61,18 +90,67 @@ def create_tables():
     """
     cfg_ehv = egon.data.config.datasets()['ehv_substation']
     cfg_hvmv = egon.data.config.datasets()['hvmv_substation']
+    cfg_ehv_voronoi = egon.data.config.datasets()['ehv_substation_voronoi']
+    cfg_hvmv_voronoi = egon.data.config.datasets()['hvmv_substation_voronoi']
     db.execute_sql(
         f"CREATE SCHEMA IF NOT EXISTS {cfg_ehv['processed']['schema']};")
 
-    db.execute_sql(f"""DROP TABLE IF EXISTS {cfg_ehv['processed']['schema']}.
-                   {cfg_ehv['processed']['table']} CASCADE;""")
+    # Drop tables
+    db.execute_sql(
+        f"""DROP TABLE IF EXISTS
+            {cfg_ehv['processed']['schema']}.
+            {cfg_ehv['processed']['table']} CASCADE;"""
+    )
 
-    db.execute_sql(f"""DROP TABLE IF EXISTS {cfg_hvmv['processed']['schema']}.
-                   {cfg_hvmv['processed']['table']} CASCADE;""")
+    db.execute_sql(
+        f"""DROP TABLE IF EXISTS
+            {cfg_hvmv['processed']['schema']}.
+            {cfg_hvmv['processed']['table']} CASCADE;"""
+    )
+
+    db.execute_sql(
+        f"""DROP TABLE IF EXISTS
+            {cfg_ehv_voronoi['processed']['schema']}.
+            {cfg_ehv_voronoi['processed']['table']} CASCADE;"""
+    )
+
+    db.execute_sql(
+        f"""DROP TABLE IF EXISTS
+            {cfg_hvmv_voronoi['processed']['schema']}.
+            {cfg_hvmv_voronoi['processed']['table']} CASCADE;"""
+    )
+
+    # Drop sequences
+    db.execute_sql(
+        f"""DROP SEQUENCE IF EXISTS
+            {cfg_ehv_voronoi['processed']['schema']}.
+            {cfg_ehv_voronoi['processed']['table']}_id_seq CASCADE;"""
+    )
+
+    db.execute_sql(
+        f"""DROP SEQUENCE IF EXISTS
+            {cfg_hvmv_voronoi['processed']['schema']}.
+            {cfg_hvmv_voronoi['processed']['table']}_id_seq CASCADE;"""
+    )
+
+    db.execute_sql(
+        f"""DROP SEQUENCE IF EXISTS
+            {cfg_hvmv['processed']['schema']}.
+            {cfg_hvmv['processed']['table']}_subst_id_seq CASCADE;"""
+    )
+
+    db.execute_sql(
+        f"""DROP SEQUENCE IF EXISTS
+            {cfg_ehv['processed']['schema']}.
+            {cfg_ehv['processed']['table']}_subst_id_seq CASCADE;"""
+    )
 
     engine = db.engine()
     EgonEhvSubstation.__table__.create(bind=engine, checkfirst=True)
     EgonHvmvSubstation.__table__.create(bind=engine, checkfirst=True)
+    EgonEhvSubstationVoronoi.__table__.create(bind=engine, checkfirst=True)
+    EgonHvmvSubstationVoronoi.__table__.create(bind=engine, checkfirst=True)
+
 
 def create_sql_functions():
     """Defines Postgresql functions needed to extract substation from osm
@@ -162,3 +240,73 @@ def create_sql_functions():
         COST 100;
         """
         )
+
+def create_voronoi():
+    '''
+    Creates voronoi polygons for hvmv and ehv substations
+
+    Returns
+    -------
+    None.
+
+    '''
+
+    substation_list = ['hvmv_substation', 'ehv_substation']
+
+    for substation in substation_list:
+        schema = egon.data.config.datasets()[substation]['processed']['schema']
+        substation_table = egon.data.config.datasets()[substation]['processed']['table']
+        voronoi_table = egon.data.config.datasets()[substation + '_voronoi']['processed']['table']
+        view = 'grid.egon_voronoi_no_borders'
+        boundary = 'boundaries.vg250_sta_union'
+
+
+        # Create view for Voronoi polygons without taking borders into account
+        db.execute_sql(
+            f"DROP VIEW IF EXISTS {schema}.egon_voronoi_no_borders CASCADE;"
+                       )
+
+        db.execute_sql(
+            f"""
+            CREATE VIEW {view} AS
+               SELECT (ST_Dump(ST_VoronoiPolygons(ST_collect(a.point)))).geom
+               FROM {schema}.{substation_table} a;
+            """
+            )
+
+        # Clip Voronoi with boundaries
+        db.execute_sql(
+            f"""
+            INSERT INTO {schema}.{voronoi_table} (geom)
+            (SELECT ST_Multi(ST_Intersection(
+                ST_Transform(a.geometry, 4326), b.geom)) AS geom
+             FROM {boundary} a
+             CROSS JOIN {view} b);
+            """
+            )
+
+        # Assign substation id as foreign key
+        db.execute_sql(
+            f"""
+            UPDATE {schema}.{voronoi_table} AS t1
+                SET  	subst_id = t2.subst_id
+	            FROM	(SELECT	voi.id AS id,
+			                sub.subst_id ::integer AS subst_id
+		            FROM	{schema}.{voronoi_table} AS voi,
+			                {schema}.{substation_table} AS sub
+		            WHERE  	voi.geom && sub.point AND
+			                ST_CONTAINS(voi.geom,sub.point)
+		           GROUP BY voi.id,sub.subst_id
+		           )AS t2
+	            WHERE  	t1.id = t2.id;
+            """
+            )
+
+        db.execute_sql(
+            f"""
+            CREATE INDEX  	{voronoi_table}_idx
+                ON          {schema}.{voronoi_table} USING gist (geom);
+            """
+            )
+
+        db.execute_sql(f"DROP VIEW IF EXISTS {view} CASCADE;")