+++ src/egon/data/processing/zensus_vg250/zensus_population_inside_germany.py
@@ -138,6 +138,9 @@ def inside_germany():
     engine_local_db = db.engine()
 
     # Create new table
+    DestatisZensusPopulationPerHaInsideGermany.__table__.drop(
+        bind=engine_local_db, checkfirst=True
+    )
     DestatisZensusPopulationPerHaInsideGermany.__table__.create(
         bind=engine_local_db, checkfirst=True
     )
@@ -184,6 +187,7 @@ def population_in_municipalities():
     """
 
     engine_local_db = db.engine()
+    Vg250GemPopulation.__table__.drop(bind=engine_local_db, checkfirst=True)
     Vg250GemPopulation.__table__.create(bind=engine_local_db, checkfirst=True)
 
     srid = 3035